datavault_blog_post_series/article_3/code_samples/requirements.txt
# DataVault Enterprise Implementation Requirements
# Article 3: Scaling Intelligence Across the Enterprise

# Core dependencies
aiohttp==3.9.1
python-dotenv==1.0.0
asyncio-mqtt==0.16.2

# Nuclia SDK (if available in environment)
# nuclia==4.9.8

# Database and caching
aiosqlite==0.19.0

# Data processing
pandas==2.1.4
numpy==1.26.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Monitoring and metrics
prometheus-client==0.19.0

# Security
cryptography==41.0.7
pyjwt==2.8.0